initial federated architecture
