+++ packages/camera/camera/example/pubspec.yaml
@@ -9,7 +9,7 @@ dependencies:
     sdk: flutter
   video_player: ^0.10.0
   integration_test:
-    path: ../../integration_test
+    path: ../../../integration_test
 
 dev_dependencies:
   flutter_test:
