--- packages/camera/camera/lib/new/camera.dart
@@ -1,10 +1,0 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-export 'src/camera_controller.dart';
-export 'src/camera_testing.dart';
-export 'src/common/camera_interface.dart';
-export 'src/common/native_texture.dart';
-export 'src/support_android/camera.dart';
-export 'src/support_android/camera_info.dart';
