+++ packages/camera/camera/pubspec.yaml
@@ -9,6 +9,8 @@ homepage: https://github.com/flutter/plugins/tree/master/packages/camera
 dependencies:
   flutter:
     sdk: flutter
+  camera_platform_interface:
+   path: ../camera_platform_interface
 
 dev_dependencies:
   path_provider: ^0.5.0
