+++ packages/camera/camera_platform_interface/lib/camera_image.dart
@@ -2,7 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-part of 'camera.dart';
+import 'dart:typed_data';
+
+import 'package:flutter/foundation.dart';
 
 /// A single color plane of image data.
 ///
@@ -113,7 +115,7 @@ ImageFormatGroup _asImageFormatGroup(dynamic rawFormat) {
 /// Although not all image formats are planar on iOS, we treat 1-dimensional
 /// images as single planar images.
 class CameraImage {
-  CameraImage._fromPlatformData(Map<dynamic, dynamic> data)
+  CameraImage.fromPlatformData(Map<dynamic, dynamic> data)
       : format = ImageFormat._fromPlatformData(data['format']),
         height = data['height'],
         width = data['width'],
