+++ packages/camera/camera_platform_interface/lib/camera_platform_interface.dart
@@ -1,0 +1,82 @@
+// Copyright 2017 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'dart:ui';
+
+import 'package:plugin_platform_interface/plugin_platform_interface.dart';
+
+import 'method_channel_camera.dart';
+
+/// The interface that implementations of camera must implement.
+///
+/// Platform implementations should extend this class rather than implement it as `camera`
+/// does not consider newly added methods to be breaking changes. Extending this class
+/// (using `extends`) ensures that the subclass will get the default implementation, while
+/// platform implementations that `implements` this interface will be broken by newly added
+/// [CameraPlatform] methods.
+abstract class CameraPlatform extends PlatformInterface {
+  /// Constructs a UrlLauncherPlatform.
+  CameraPlatform() : super(token: _token);
+
+  static final Object _token = Object();
+
+  static CameraPlatform _instance = MethodChannelCamera();
+
+  /// The default instance of [CameraPlatform] to use.
+  ///
+  /// Defaults to [MethodChannelCamera].
+  static CameraPlatform get instance => _instance;
+
+  /// Platform-specific plugins should set this with their own platform-specific
+  /// class that extends [UrlLauncherPlatform] when they register themselves.
+  static set instance(CameraPlatform instance) {
+    PlatformInterface.verifyToken(instance, _token);
+    _instance = instance;
+  }
+
+
+  /// Completes with a list of available cameras.
+  ///
+  /// May throw a [CameraException].
+  Future<List<CameraDescription>> availableCameras() async {
+    throw UnimplementedError('availableCameras() has not been implemented.');
+  }
+}
+
+enum CameraLensDirection { front, back, external }
+
+/// TODO: docs
+class CameraDescription {
+  /// TODO: docs
+  CameraDescription({this.name, this.lensDirection, this.sensorOrientation});
+
+  final String name;
+  final CameraLensDirection lensDirection;
+
+  /// Clockwise angle through which the output image needs to be rotated to be upright on the device screen in its native orientation.
+  ///
+  /// **Range of valid values:**
+  /// 0, 90, 180, 270
+  ///
+  /// On Android, also defines the direction of rolling shutter readout, which
+  /// is from top to bottom in the sensor's coordinate system.
+  final int sensorOrientation;
+
+  @override
+  bool operator ==(Object o) {
+    return o is CameraDescription &&
+        o.name == name &&
+        o.lensDirection == lensDirection;
+  }
+
+  @override
+  int get hashCode {
+    return hashValues(name, lensDirection);
+  }
+
+  @override
+  String toString() {
+    return '$runtimeType($name, $lensDirection, $sensorOrientation)';
+  }
+}
