+++ packages/camera/camera_platform_interface/pubspec.yaml
@@ -1,0 +1,22 @@
+name: camera_platform_interface
+description: A common platform interface for the camera plugin.
+homepage: https://github.com/flutter/plugins/tree/master/packages/camera/camera_platform_interface
+# NOTE: We strongly prefer non-breaking changes, even at the expense of a
+# less-clean API. See https://flutter.dev/go/platform-interface-breaking-changes
+version: 1.0.0
+
+dependencies:
+  flutter:
+    sdk: flutter
+  meta: ^1.0.5
+  plugin_platform_interface: ^1.0.1
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  mockito: ^4.1.1
+  pedantic: ^1.8.0
+
+environment:
+  sdk: ">=2.8.0 <3.0.0"
+  flutter: ">=1.10.0 <2.0.0"
